# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API Documentation",
      "version": "1.0.0"
    },
    "paths": {
...
      "/user/register": {
        "post": {
          "summary": "Register a new user",
          "tags": ["Authentication"],
          "operationId": "registerUser",
          "requestBody": {
            "description": "User registration details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  },
                  "required": ["email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User created",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean"
                      },
                      "message": {
                        "type": "string",
                        "default":"Request body incomplete, both email and password are required"
                      }
                    }
                  }
                }
              }
            },
            "409": {
              "description": "User already exists",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean"
                      },
                      "message": {
                        "type": "string",
                        "default":"User already exists"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "default":"Internal server error"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/user/login": {
        "post": {
          "summary": "User login",
          "tags": ["Authentication"],
          "operationId": "loginUser",
          "requestBody": {
            "description": "User login credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    },
                    "longExpiry": {
                      "type": "boolean",
                      "default": false
                    }
                  },
                  "required": ["email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful login",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "bearerToken": {
                        "type": "object",
                        "properties": {
                          "token": {
                            "type": "string"
                          },
                          "token_type": {
                            "type": "string",
                            "enum": ["Bearer"]
                          },
                          "expires_in": {
                            "type": "string"
                          }
                        }
                      },
                      "refreshToken": {
                        "type": "object",
                        "properties": {
                          "token": {
                            "type": "string"
                          },
                          "token_type": {
                            "type": "string",
                            "enum": ["Refresh"]
                          },
                          "expires_in": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"Request body incomplete, both email and password are required"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"Incorrect email or password"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "default":"Internal server error"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/user/refresh": {
        "post": {
          "summary": "Refresh bearer token",
          "tags": ["Authentication"],
          "operationId": "refreshBearerToken",
          "requestBody": {
            "description": "Refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "refreshToken": {
                      "type": "string"
                    }
                  },
                  "required": ["refreshToken"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful token refresh",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "bearerToken": {
                        "type": "object",
                        "properties": {
                          "token": {
                            "type": "string"
                          },
                          "token_type": {
                            "type": "string",
                            "enum": ["Bearer"]
                          },
                          "expires_in": {
                            "type": "string"
                          }
                        }
                      },
                      "refreshToken": {
                        "type": "object",
                        "properties": {
                          "token": {
                            "type": "string"
                          },
                          "token_type": {
                            "type": "string",
                            "enum": ["Refresh"]
                          },
                          "expires_in": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"Request body incomplete, refresh token required"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"Invalid refresh token"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "default":"Internal server error"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "GET /user/{email}/profile": {
        "get": {
          "summary": "Get user profile",
          "tags": ["Authentication"],
          "operationId": "getUser",
          "parameters": [
            {
              "name": "email",
              "in": "path",
              "description": "User email",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User profile retrieved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "email": {
                        "type": "string"
                      },
                      "firstName": {
                        "type": "string"
                      },
                      "lastName": {
                        "type": "string"
                      },
                      "dob": {
                        "type": "string",
                        "format": "date"
                      },
                      "address": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "User not found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"User not found"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"Internal server error"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "PUT /user/{email}/profile": {
        "put": {
          "summary": "Update user profile",
          "tags": ["Authentication"],
          "operationId": "putUser",
          "parameters": [
            {
              "name": "email",
              "in": "path",
              "description": "User email",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "dob": {
                      "type": "string",
                      "format": "date"
                    },
                    "address": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "User profile updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "email": {
                        "type": "string"
                      },
                      "firstName": {
                        "type": "string"
                      },
                      "lastName": {
                        "type": "string"
                      },
                      "dob": {
                        "type": "string",
                        "format": "date"
                      },
                      "address": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"Authorization header ('Bearer token') not found"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "User not found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"User not found"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"Internal server error"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }            
    }
  }